Optimized lazy loading in rmw-shell template
