+++ packages/base-shell/package.json
@@ -1,6 +1,6 @@
 {
   "name": "base-shell",
-  "version": "1.2.13",
+  "version": "1.2.14",
   "description": "base-shell React component",
   "main": "lib/index.js",
   "module": "es/index.js",
