+++ packages/rmw-shell/cra-template-rmw/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cra-template-rmw",
-  "version": "2.0.10",
+  "version": "2.1.0",
   "keywords": [
     "react",
     "create-react-app",
