+++ packages/rmw-shell/cra-template-rmw/template.json
@@ -1,52 +1,41 @@
 {
-  "dependencies": {
-    "rmw-shell":"8.x",
-    "@formatjs/intl-relativetimeformat": "7.x",
-    "@material-ui/core": "4.x",
-    "@material-ui/icons": "4.x",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@material-ui/pickers": "3.x",
-    "base-shell": "1.x",
-    "chart.js": "2.x",
-    "final-form": "4.x",
-    "firebase": "7.x",
-    "firebaseui": "4.x",
-    "github-markdown-css": "4.x",
-    "intl": "1.x",
-    "intl-locales-supported": "1.x",
-    "material-ui-shell": "1.x",
-    "moment": "2.x",
-    "mui-rff": "2.x",
-    "notistack": "1.x",
-    "react": "17.x",
-    "react-chartjs-2": "2.x",
-    "react-countup": "4.x",
-    "react-custom-scrollbars": "4.x",
-    "react-dom": "17.x",
-    "react-easy-crop": "3.x",
-    "react-final-form": "6.x",
-    "react-helmet": "6.x",
-    "react-intl": "5.x",
-    "react-ios-pwa-prompt": "1.x",
-    "react-markdown": "5.x",
-    "react-router-dom": "5.x",
-    "react-virtualized-auto-sizer": "1.x",
-    "react-window": "1.x",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "web-vitals": "^0.2.4",
-    "workbox-background-sync": "^5.1.3",
-    "workbox-broadcast-update": "^5.1.3",
-    "workbox-cacheable-response": "^5.1.3",
-    "workbox-core": "^5.1.3",
-    "workbox-expiration": "^5.1.3",
-    "workbox-google-analytics": "^5.1.3",
-    "workbox-navigation-preload": "^5.1.3",
-    "workbox-precaching": "^5.1.3",
-    "workbox-range-requests": "^5.1.3",
-    "workbox-routing": "^5.1.3",
-    "workbox-strategies": "^5.1.3",
-    "workbox-streams": "^5.1.3"
+  "package": {
+    "dependencies": {
+      "rmw-shell":"8.x",
+      "@formatjs/intl-relativetimeformat": "7.x",
+      "@material-ui/core": "4.x",
+      "@material-ui/icons": "4.x",
+      "@material-ui/lab": "^4.0.0-alpha.56",
+      "@material-ui/pickers": "3.x",
+      "base-shell": "1.x",
+      "chart.js": "2.x",
+      "final-form": "4.x",
+      "firebase": "7.x",
+      "firebaseui": "4.x",
+      "github-markdown-css": "4.x",
+      "intl": "1.x",
+      "intl-locales-supported": "1.x",
+      "material-ui-shell": "1.x",
+      "moment": "2.x",
+      "mui-rff": "2.x",
+      "notistack": "1.x",
+      "react": "17.x",
+      "react-chartjs-2": "^2.x",
+      "react-countup": "4.x",
+      "react-custom-scrollbars": "4.x",
+      "react-dom": "17.x",
+      "react-easy-crop": "3.x",
+      "react-final-form": "6.x",
+      "react-helmet": "6.x",
+      "react-intl": "5.x",
+      "react-ios-pwa-prompt": "1.x",
+      "react-markdown": "5.x",
+      "react-router-dom": "5.x",
+      "react-virtualized-auto-sizer": "1.x",
+      "react-window": "1.x"
+    },
+    "eslintConfig": {
+      "extends": ["react-app", "react-app/jest"]
+    }
   }
 }
