--- packages/rmw-shell/cra-template-rmw/template/package.json
@@ -1,77 +1,0 @@
-{
-  "name": "rmw-template",
-  "version": "0.1.2",
-  "private": true,
-  "dependencies": {
-    "@formatjs/intl-relativetimeformat": "^7.3.1",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@material-ui/pickers": "^3.2.10",
-    "base-shell": "^1.2.11",
-    "chart.js": "^2.9.4",
-    "eslint-loader": "^4.0.2",
-    "final-form": "^4.20.1",
-    "firebase": "^7.24.0",
-    "firebaseui": "^4.7.0",
-    "github-markdown-css": "^4.0.0",
-    "intl": "^1.2.5",
-    "intl-locales-supported": "^1.8.12",
-    "material-ui-shell": "^1.1.19",
-    "moment": "^2.29.1",
-    "mui-rff": "^2.5.6",
-    "notistack": "^1.0.1",
-    "nwb": "^0.25.2",
-    "rmw-shell": "^8.0.0",
-    "raw-loader": "^4.0.2",
-    "react": "^17.0.1",
-    "react-chartjs-2": "^2.10.0",
-    "react-countup": "^4.3.3",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.1",
-    "react-easy-crop": "^3.2.2",
-    "react-final-form": "^6.5.2",
-    "react-helmet": "^6.1.0",
-    "react-intl": "^5.8.8",
-    "react-ios-pwa-prompt": "^1.8.0",
-    "react-markdown": "^5.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.0",
-    "react-virtualized-auto-sizer": "^1.0.2",
-    "react-window": "^1.8.5"
-  },
-  "scripts": {
-    "check": "npm-check -s -u",
-    "analyze": "source-map-explorer build/static/js/main.*",
-    "start": "react-scripts start",
-    "standard": "standard src/**/*.js ",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom --passWithNoTests",
-    "eject": "react-scripts eject"
-  },
-  "author": "Tarik Huber",
-  "homepage": "https://www.react-most-wanted.com",
-  "license": "MIT",
-  "repository": "https://github.com/TarikHuber/react-most-wanted",
-  "keywords": [
-    "react-most-wanted",
-    "react",
-    "redux",
-    "firebase"
-  ],
-  "prettier": {
-    "trailingComma": "es5",
-    "tabWidth": 2,
-    "semi": false,
-    "singleQuote": true
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "devDependencies": {
-    "selenium-webdriver": "^4.0.0-alpha.5"
-  }
-}
