+++ packages/rmw-shell/cra-template-rmw/template/src/App.js
@@ -1,10 +1,6 @@
 import React, { Component } from 'react'
 import App from 'base-shell/lib'
-import MUIConfig from 'rmw-shell/lib/config'
-import merge from 'base-shell/lib/utils/config'
-import _config from './config'
-
-const config = merge(MUIConfig, _config)
+import config from './config'
 
 export default class Demo extends Component {
   render() {
