+++ packages/rmw-shell/cra-template-rmw/template/src/config/getDefaultRoutes.js
@@ -1,25 +1,25 @@
-import AuthorizedRoute from 'base-shell/lib/components/AuthorizedRoute/AuthorizedRoute'
-import React, { lazy } from 'react'
-import UnauthorizedRoute from 'base-shell/lib/components/UnauthorizedRoute/UnauthorizedRoute'
-import { Route } from 'react-router-dom'
+import AuthorizedRoute from "base-shell/lib/components/AuthorizedRoute/AuthorizedRoute";
+import React, { lazy } from "react";
+import UnauthorizedRoute from "base-shell/lib/components/UnauthorizedRoute/UnauthorizedRoute";
+import { Route } from "react-router-dom";
 
-const SignIn = lazy(() => import('../pages/SignIn/SignIn'))
-const MyAccount = lazy(() => import('../pages/MyAccount/MyAccount'))
-const Users = lazy(() => import('../pages/Users'))
-const User = lazy(() => import('../pages/Users/User'))
-const Roles = lazy(() => import('../pages/Roles'))
-const Role = lazy(() => import('../pages/Roles/Role'))
-const Chats = lazy(() => import('../pages/Chats'))
-const CreateChat = lazy(() => import('../pages/CreateChat'))
-const CreateGroup = lazy(() => import('../pages/CreateGroup'))
-const EditMembers = lazy(() => import('../pages/EditMembers'))
-const EditAdmins = lazy(() => import('../pages/EditAdmins'))
+const SignIn = lazy(() => import("rmw-shell/lib/pages/SignIn/SignIn"));
+const MyAccount = lazy(() => import("rmw-shell/lib/pages/MyAccount/MyAccount"));
+const Users = lazy(() => import("rmw-shell/lib/pages/Users"));
+const User = lazy(() => import("rmw-shell/lib/pages/Users/User"));
+const Roles = lazy(() => import("rmw-shell/lib/pages/Roles"));
+const Role = lazy(() => import("rmw-shell/lib/pages/Roles/Role"));
+const Chats = lazy(() => import("rmw-shell/lib/pages/Chats"));
+const CreateChat = lazy(() => import("rmw-shell/lib/pages/CreateChat"));
+const CreateGroup = lazy(() => import("rmw-shell/lib/pages/CreateGroup"));
+const EditMembers = lazy(() => import("rmw-shell/lib/pages/EditMembers"));
+const EditAdmins = lazy(() => import("rmw-shell/lib/pages/EditAdmins"));
 
 const getDefaultRoutes = (appConfig) => {
   return [
     <UnauthorizedRoute
       path="/signin"
-      redirectTo={appConfig?.auth?.redirectTo || '/'}
+      redirectTo={appConfig?.auth?.redirectTo || "/"}
       exact
       component={SignIn}
     />,
@@ -39,7 +39,7 @@ const getDefaultRoutes = (appConfig) => {
     <AuthorizedRoute path="/users/:uid/:tab" exact component={User} />,
 
     <Route component={appConfig?.pages?.PageNotFound} />,
-  ]
-}
+  ];
+};
 
-export default getDefaultRoutes
+export default getDefaultRoutes;
