+++ packages/rmw-shell/nwb.config.js
@@ -10,11 +10,5 @@ module.exports = {
     aliases: {
       'rmw-shell/lib': path.resolve('src'),
     },
-    extra: {
-      // Adding an extra rule which isn't managed by nwb
-      module: {
-        rules: [{ test: /\.md$/, loader: 'raw-loader' }],
-      },
-    },
   },
 }
