+++ packages/rmw-shell/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rmw-shell",
-  "version": "8.0.5",
+  "version": "8.1.0",
   "description": "React template with Material UI, Firebase, routing...",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -39,7 +39,7 @@
     "mui-rff": "2.x",
     "notistack": "1.x",
     "react": "17.x",
-    "react-chartjs-2": "2.x",
+    "react-chartjs-2": "^2.x",
     "react-countup": "4.x",
     "react-custom-scrollbars": "4.x",
     "react-dom": "17.x",
@@ -54,12 +54,12 @@
     "react-window": "1.x"
   },
   "devDependencies": {
-    "@formatjs/intl-relativetimeformat": "^7.3.2",
+    "@formatjs/intl-relativetimeformat": "^7.3.3",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
     "@material-ui/pickers": "^3.2.10",
-    "base-shell": "^1.2.13",
+    "base-shell": "^1.2.14",
     "chart.js": "^2.9.4",
     "eslint-loader": "^4.0.2",
     "final-form": "^4.20.1",
@@ -68,12 +68,11 @@
     "github-markdown-css": "^4.0.0",
     "intl": "^1.2.5",
     "intl-locales-supported": "^1.8.12",
-    "material-ui-shell": "^1.1.22",
+    "material-ui-shell": "^1.1.24",
     "moment": "^2.29.1",
     "mui-rff": "^2.5.6",
     "notistack": "^1.0.1",
     "nwb": "^0.25.2",
-    "raw-loader": "^4.0.2",
     "react": "^17.0.1",
     "react-chartjs-2": "^2.10.0",
     "react-countup": "^4.3.3",
@@ -82,7 +81,7 @@
     "react-easy-crop": "^3.3.0",
     "react-final-form": "^6.5.2",
     "react-helmet": "^6.1.0",
-    "react-intl": "^5.8.9",
+    "react-intl": "^5.9.2",
     "react-ios-pwa-prompt": "^1.8.0",
     "react-markdown": "^5.0.2",
     "react-router-dom": "^5.2.0",
