--- packages/rmw-shell/src/config/grants.js
@@ -1,3 +1,0 @@
-export const grants = []
-
-export default grants
