--- packages/rmw-shell/src/config/index.js
@@ -1,1 +1,0 @@
-export { default } from './config.js'
